Mainly work on material serialization
Plus a plethoria of other changes
added detete / route, to be blinded in prod I assume

+++ event_horizon/core/platform/desktop/webclient_desktop.cpp
@@ -70,7 +70,7 @@ namespace Http {
 
         res.uri = url.toString();
         res.statusCode = response->get_status_code();
-        res.length = response->get_header( "Content-Length", 0 );
+        res.length = static_cast<uint64_t>(response->get_header( "Content-Length", 0 ));
         res.contentType = response->get_header( "Content-Type", "application/json" );
         res.ETag = response->get_header( "ETag", "" );
         res.lastModified = response->get_header( "Content-Last-Modified", "" );
